+++ frontend/mirror-frontend/src/components/App.css
