+++ frontend/mirror-frontend/src/components/flask_weather.py
