+++ frontend/mirror-frontend/src/svg.d.ts
